Split Turkey specs one class per file and relocate the command specs
